Refactoring of merge_enum code to its own file
