+++ apollo-federation/src/merger/merge_union.rs
@@ -113,7 +113,7 @@ mod tests {
     use apollo_compiler::schema::ObjectType;
 
     use super::*;
-    use crate::merger::merge::tests::create_test_merger;
+    use crate::merger::merge_enum::tests::create_test_merger;
     use crate::schema::position::ObjectTypeDefinitionPosition;
 
     // Helper function to create a union type for testing
