+++ apollo-federation/src/merger/mod.rs
@@ -3,4 +3,5 @@ mod error_reporter;
 mod hints;
 #[path = "merger.rs"]
 mod merge;
+mod merge_enum;
 mod merge_union;
